add direcotories to manage these scripts
