+++ src/modules/utils/commit-info.js
@@ -1,0 +1,1 @@
+export const COMMIT_ID = '1234e3ceb37a'
